requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
requests
python-dotenv
